int to parseFloat

+++ js/main_js.js
@@ -2,27 +2,27 @@ function convert(){
     function calc(base) {
     let decimal = document.getElementById("decimal").value;
     let modi="";
-    while (parseInt(decimal)/base > 0) {
+    while (parseFloat(decimal)/base > 0) {
             if (base === 16) {
-                if (parseInt(decimal) % base < 10) {
-                    modi = modi + (parseInt(decimal) % base );
-                } else if (parseInt(decimal) % base === 10) {
+                if (parseFloat(decimal) % base < 10) {
+                    modi = modi + (parseFloat(decimal) % base );
+                } else if (parseFloat(decimal) % base === 10) {
                     modi = modi + "A";
-                }else if (parseInt(decimal) % base === 11) {
+                }else if (parseFloat(decimal) % base === 11) {
                     modi = modi + "B";
-                }else if (parseInt(decimal) % base === 12) {
+                }else if (parseFloat(decimal) % base === 12) {
                     modi = modi + "C";
-                }else if (parseInt(decimal) % base === 13) {
+                }else if (parseFloat(decimal) % base === 13) {
                     modi = modi + "D";
-                }else if (parseInt(decimal) % base === 14) {
+                }else if (parseFloat(decimal) % base === 14) {
                     modi = modi + "E";
-                }else if (parseInt(decimal) % base === 15) {
+                }else if (parseFloat(decimal) % base === 15) {
                     modi = modi + "F";
                 }
             } else {
-                modi = modi + (parseInt(decimal) % base );
+                modi = modi + (parseFloat(decimal) % base );
             }
-        decimal = parseInt(decimal)/base;
+        decimal = parseFloat(decimal)/base;
     }
     function reverseString(str) {
         return str.split("").reverse().join("");
